fix 17.08.2021

+++ src/main/java/org/hostel/dto/CategoryDto.java
@@ -2,14 +2,14 @@ package org.hostel.dto;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
-import org.hostel.domains.Category;
-import org.hostel.domains.CategoryName;
+import org.hostel.domain.Category;
+import org.hostel.domain.CategoryName;
 
 @Data
 @RequiredArgsConstructor
 public class CategoryDto {
 
-        private int id;
+        private long id;
         private CategoryName categoryName;
         private String description;
